requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
APScheduler==3.11.0
click==8.1.7
dnspython==2.7.0
fastapi==0.95.2
geographiclib==2.0
geopy==2.3.0
h11==0.14.0
idna==3.10
pydantic==1.10.7
pydantic_core==2.23.4
pymongo==4.5.0
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.27.0
typing_extensions==4.12.2
tzlocal==5.2
uvicorn==0.22.0
